# aspnetcore-server-generated/src/IO.Swagger/Controllers/OperationsControlApi.cs
/*
 * Service-Oriented Vehicle Diagnostic API
 *
 * Definition of ASAM SOVD API
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class OperationsControlApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method returns all available operations defined for the given entity.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="includeProximityProof">Specifies whether operations which require a proximity proof are returned or not. The default is true.</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <response code="200">Response: Retrieve list of all available operations from an entity</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/operations")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdOperationsGet")]
        public virtual IActionResult EntityCollectionEntityIdOperationsGet([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromQuery]Object includeProximityProof, [FromQuery]Object includeSchema)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method terminates the execution of the operation and removes it, and the status of the operation can no longer be requested.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="operationId">Identifier for the operation</param>
        /// <param name="executionId">Identifier for the operation execution</param>
        /// <response code="204">The request was successful.</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpDelete]
        [Route("/{entity-collection}/{entity-id}/operations/{operation-id}/executions/{execution-id}")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdOperationsOperationIdExecutionsExecutionIdDelete")]
        public virtual IActionResult EntityCollectionEntityIdOperationsOperationIdExecutionsExecutionIdDelete([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object operationId, [FromRoute][Required]Object executionId)
        { 
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(204);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method returns the current status of the operation execution.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="operationId">Identifier for the operation</param>
        /// <param name="executionId">Identifier for the operation execution</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <response code="200">Response: Get the status of an operation execution</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/operations/{operation-id}/executions/{execution-id}")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdOperationsOperationIdExecutionsExecutionIdGet")]
        public virtual IActionResult EntityCollectionEntityIdOperationsOperationIdExecutionsExecutionIdGet([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object operationId, [FromRoute][Required]Object executionId, [FromQuery]Object includeSchema)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method executes the given capability on the provided operation.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="operationId">Identifier for the operation</param>
        /// <param name="executionId">Identifier for the operation execution</param>
        /// <response code="202">Response: Start execution of an operation - asynchronous execution</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpPut]
        [Route("/{entity-collection}/{entity-id}/operations/{operation-id}/executions/{execution-id}")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdOperationsOperationIdExecutionsExecutionIdPut")]
        public virtual IActionResult EntityCollectionEntityIdOperationsOperationIdExecutionsExecutionIdPut([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object operationId, [FromRoute][Required]Object executionId)
        { 
            //TODO: Uncomment the next line to return response 202 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(202);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method returns currently existing executions of the given operation.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="operationId">Identifier for the operation</param>
        /// <response code="200">Response: Get executions of an operation</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/operations/{operation-id}/executions")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdOperationsOperationIdExecutionsGet")]
        public virtual IActionResult EntityCollectionEntityIdOperationsOperationIdExecutionsGet([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object operationId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Start Execution of an Operation An operation may support multiple executions in parallel. If this is not supported, this method will return an error if an operation with the same operation-id is already being executed. A successfully started operation is added to the executions collection if it is executed asynchronously. After the asynchronous execution has been finished, the execution resource remains existent and an SOVD client can receive the status. However, the status of the execution cannot be queried indefinitely as an SOVD server implementation may remove the resource, e.g., upon SOVD server restart.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="operationId">Identifier for the operation</param>
        /// <response code="200">Response: Read Configurations</response>
        /// <response code="202">Response: Start execution of an operation - asynchronous execution</response>
        /// <response code="409">An unexpected request occurred</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpPost]
        [Route("/{entity-collection}/{entity-id}/operations/{operation-id}/executions")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdOperationsOperationIdExecutionsPost")]
        public virtual IActionResult EntityCollectionEntityIdOperationsOperationIdExecutionsPost([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object operationId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 202 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(202);

            //TODO: Uncomment the next line to return response 409 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(409);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method returns all the details on the specified operation of the given entity.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="operationId">Identifier for the operation</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <response code="200">Response: Get details of single operation</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/operations/{operation-id}")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdOperationsOperationIdGet")]
        public virtual IActionResult EntityCollectionEntityIdOperationsOperationIdGet([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object operationId, [FromQuery]Object includeSchema)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }
    }
}