# aspnetcore-server-generated/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Definition of ASAM SOVD API",
    "version" : "1.0.0",
    "title" : "Service-Oriented Vehicle Diagnostic API",
    "contact" : {
      "url" : "https://www.asam.net/project-detail/sovd-service-oriented-vehicle-diagnostics/"
    }
  },
...
  "paths" : {
...
    "/{entity-collection}/{entity-id}/operations" : {
      "get" : {
        "tags" : [ "operations-control" ],
        "description" : "This method returns all available operations defined for the given entity.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "include-proximity-proof",
          "in" : "query",
          "description" : "Specifies whether operations which require a proximity proof are returned or not. The default is true.",
          "required" : false
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Retrieve list of all available operations from an entity"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
    "/{entity-collection}/{entity-id}/operations/{operation-id}" : {
      "get" : {
        "tags" : [ "operations-control" ],
        "description" : "This method returns all the details on the specified operation of the given entity.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "operation-id",
          "in" : "path",
          "description" : "Identifier for the operation",
          "required" : true
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Get details of single operation"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        },
        "x-sovd-asynchronous-execution" : true
      }
    },
    "/{entity-collection}/{entity-id}/operations/{operation-id}/executions" : {
      "get" : {
        "tags" : [ "operations-control" ],
        "description" : "This method returns currently existing executions of the given operation.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "operation-id",
          "in" : "path",
          "description" : "Identifier for the operation",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Get executions of an operation"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      },
      "post" : {
        "tags" : [ "operations-control" ],
        "description" : "Start Execution of an Operation An operation may support multiple executions in parallel. If this is not supported, this method will return an error if an operation with the same operation-id is already being executed. A successfully started operation is added to the executions collection if it is executed asynchronously. After the asynchronous execution has been finished, the execution resource remains existent and an SOVD client can receive the status. However, the status of the execution cannot be queried indefinitely as an SOVD server implementation may remove the resource, e.g., upon SOVD server restart.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "operation-id",
          "in" : "path",
          "description" : "Identifier for the operation",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Read Configurations"
          },
          "202" : {
            "description" : "Response: Start execution of an operation - asynchronous execution",
            "headers" : {
              "Location" : {
                "description" : "The response for status code 202 includes the following header to redirect the SOVD client for requesting the status: Location: /{entity-path}/operations/{operation-id}/executions/{execution-id}"
              }
            }
          },
          "409" : {
            "description" : "An unexpected request occurred"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
    "/{entity-collection}/{entity-id}/operations/{operation-id}/executions/{execution-id}" : {
      "get" : {
        "tags" : [ "operations-control" ],
        "description" : "This method returns the current status of the operation execution.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "operation-id",
          "in" : "path",
          "description" : "Identifier for the operation",
          "required" : true
        }, {
          "name" : "execution-id",
          "in" : "path",
          "description" : "Identifier for the operation execution",
          "required" : true
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Get the status of an operation execution"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      },
      "put" : {
        "tags" : [ "operations-control" ],
        "description" : "This method executes the given capability on the provided operation.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "operation-id",
          "in" : "path",
          "description" : "Identifier for the operation",
          "required" : true
        }, {
          "name" : "execution-id",
          "in" : "path",
          "description" : "Identifier for the operation execution",
          "required" : true
        } ],
        "responses" : {
          "202" : {
            "description" : "Response: Start execution of an operation - asynchronous execution",
            "headers" : {
              "Location" : {
                "description" : "The response for status code 202 includes the following header to redirect the SOVD client for requesting the status: Location: /{entity-path}/operations/{operation-id}/executions/{execution-id}"
              }
            }
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        },
        "x-sovd-capabilities" : [ "execute", "freeze", "reset" ]
      },
      "delete" : {
        "tags" : [ "operations-control" ],
        "description" : "This method terminates the execution of the operation and removes it, and the status of the operation can no longer be requested.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "operation-id",
          "in" : "path",
          "description" : "Identifier for the operation",
          "required" : true
        }, {
          "name" : "execution-id",
          "in" : "path",
          "description" : "Identifier for the operation execution",
          "required" : true
        } ],
        "responses" : {
          "204" : {
            "description" : "The request was successful."
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
...
  },
...
}